ml/pipeline/fix_stage1_engagement_calibration.py
"""
Phase 3 Fix — Engagement_Score interval calibration

Retrains ONLY:
  - engagement_score_q10 with alpha=0.05
  - engagement_score_q90 with alpha=0.95

Keeps engagement_score_q50 unchanged.
Recomputes test interval coverage (q10..q90) in scaled space and updates:
  - artifacts/stage1/stage1_confidence_tiers.json
  - reports/phase3_stage1_report.txt (append update note)
"""

from __future__ import annotations

import json
import pickle
from pathlib import Path

import numpy as np
import pandas as pd

import lightgbm as lgb
from sklearn.metrics import r2_score


ROOT = Path(__file__).resolve().parent.parent
DATA_DIR = ROOT / "data"
ARTIFACTS_DIR = ROOT / "artifacts"
STAGE1_DIR = ARTIFACTS_DIR / "stage1"


LGBM_PARAMS = {
    "n_estimators": 500,
    "learning_rate": 0.05,
    "num_leaves": 63,
    "min_child_samples": 50,
    "subsample": 0.8,
    "colsample_bytree": 0.8,
    "random_state": 42,
    "verbosity": -1,
}


def interval_coverage(y_true: np.ndarray, q_lo: np.ndarray, q_hi: np.ndarray) -> float:
    lo = np.minimum(q_lo, q_hi)
    hi = np.maximum(q_lo, q_hi)
    return float(np.mean((y_true >= lo) & (y_true <= hi)))


def tier_for_target(r2: float, coverage: float) -> str:
    if r2 >= 0.7 and coverage >= 0.80:
        return "HIGH"
    if r2 >= 0.5 and coverage >= 0.70:
        return "MEDIUM"
    return "LOW"


def main() -> int:
    train = pd.read_csv(DATA_DIR / "train.csv")
    val = pd.read_csv(DATA_DIR / "val.csv")
    test = pd.read_csv(DATA_DIR / "test.csv")

    feature_cols_path = STAGE1_DIR / "stage1_feature_columns.json"
    with open(feature_cols_path, "r", encoding="utf-8") as f:
        feature_cols = json.load(f)

    X_train = train[feature_cols].astype(np.float32)
    X_val = val[feature_cols].astype(np.float32)
    X_test = test[feature_cols].astype(np.float32)

    y_train = train["Engagement_Score_scaled"].to_numpy()
    y_val = val["Engagement_Score_scaled"].to_numpy()
    y_test = test["Engagement_Score_scaled"].to_numpy()

    # Load existing q50 model (do NOT retrain)
    q50_path = STAGE1_DIR / "engagement_score_q50.pkl"
    with open(q50_path, "rb") as f:
        model_q50 = pickle.load(f)
    q50_pred = model_q50.predict(X_test)
    r2_q50 = float(r2_score(y_test, q50_pred))

    def train_quantile(alpha: float, out_name: str) -> np.ndarray:
        model = lgb.LGBMRegressor(
            objective="quantile",
            alpha=alpha,
            **LGBM_PARAMS,
        )
        model.fit(
            X_train,
            y_train,
            eval_set=[(X_val, y_val)],
            eval_metric="quantile",
            callbacks=[lgb.log_evaluation(period=0)],
        )
        out_path = STAGE1_DIR / out_name
        with open(out_path, "wb") as f:
            pickle.dump(model, f)
        return model.predict(X_test)

    # Retrain ONLY q10/q90 with widened interval
    q10_pred = train_quantile(alpha=0.05, out_name="engagement_score_q10.pkl")
    q90_pred = train_quantile(alpha=0.95, out_name="engagement_score_q90.pkl")

    cov = interval_coverage(y_test, q10_pred, q90_pred)

    # Update tiers.json
    tiers_path = STAGE1_DIR / "stage1_confidence_tiers.json"
    with open(tiers_path, "r", encoding="utf-8") as f:
        tiers = json.load(f)

    new_tier = tier_for_target(r2_q50, cov)
    # User override: if still < 75% after fix, force MEDIUM and proceed anyway.
    if cov < 0.75:
        new_tier = "MEDIUM"

    tiers["Engagement_Score"] = new_tier
    with open(tiers_path, "w", encoding="utf-8") as f:
        json.dump(tiers, f, indent=2)

    # Update report (append)
    report_path = ROOT / "reports" / "phase3_stage1_report.txt"
    lines = []
    lines.append("")
    lines.append("=== Engagement_Score Calibration Update ===")
    lines.append(f"R2 (q50, test): {r2_q50:.6f}")
    lines.append(f"Interval coverage (q10..q90, test): {cov*100:.2f}%")
    lines.append(f"Tier in stage1_confidence_tiers.json: {new_tier}")
    report_path.write_text(report_path.read_text(encoding="utf-8") + "\n" + "\n".join(lines), encoding="utf-8")

    print("=== UPDATED Engagement_Score ===")
    print(f"New coverage: {cov*100:.2f}%")
    print(f"New tier: {new_tier}")
    return 0


if __name__ == "__main__":
    raise SystemExit(main())

