requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
anthropic
scikit-learn
shap
numpy
pandas
